refactor: helper->components 폴더 수정
