+++ app/[locale]/academics/components/ScholarshipCreatePage.tsx
