+++ app/[locale]/academics/components/ScholarshipDetail.tsx
