+++ app/[locale]/academics/components/ScholarshipDetailEdit.tsx
