+++ app/[locale]/academics/components/ScholarshipGuideEditPageContent.tsx
