+++ app/[locale]/academics/components/ScholarshipPreview.tsx
