+++ app/[locale]/academics/components/course-changes/CourseChangesPageContent.tsx
