+++ app/[locale]/academics/components/courses/AddCourseButton.tsx
