+++ app/[locale]/academics/components/courses/AddCourseModal.tsx
