+++ app/[locale]/academics/components/courses/CourseCard.tsx
