+++ app/[locale]/academics/components/courses/CourseCards.tsx
