+++ app/[locale]/academics/components/courses/CourseDetailModal.tsx
