+++ app/[locale]/academics/components/courses/CourseEditor.tsx
