+++ app/[locale]/academics/components/courses/CourseList.tsx
