+++ app/[locale]/academics/components/courses/CourseListHeader.tsx
