+++ app/[locale]/academics/components/courses/CourseListRow.tsx
