+++ app/[locale]/academics/components/courses/CourseRow.tsx
