+++ app/[locale]/academics/components/courses/CourseToolbar.tsx
