+++ app/[locale]/academics/components/courses/useCourseEditor.tsx
