+++ app/[locale]/academics/components/guide/GuideEditPageContent.tsx
