+++ app/[locale]/academics/components/guide/GuidePageContent.tsx
