+++ app/[locale]/academics/components/timeline/Timeline.tsx
