+++ app/[locale]/academics/components/timeline/TimelineViewer.tsx
