+++ app/[locale]/academics/graduate/scholarship/[id]/page.tsx
@@ -4,7 +4,7 @@ import { Language } from '@/types/language';
 import { getMetadata } from '@/utils/metadata';
 import { graduateScholarship } from '@/utils/segmentNode';
 
-import ScholarshipDetail from '../../../helper/ScholarshipDetail';
+import ScholarshipDetail from '../../../components/ScholarshipDetail';
 
 interface ScholarshipDetailProps {
   params: { locale: Language; id: string };
