+++ app/[locale]/academics/graduate/scholarship/page.tsx
@@ -2,7 +2,7 @@ import { getScholarshipList } from '@/apis/v1/academics/scholarship';
 import { getMetadata } from '@/utils/metadata';
 import { graduateScholarship } from '@/utils/segmentNode';
 
-import ScholarshipPreview from '../../helper/ScholarshipPreview';
+import ScholarshipPreview from '../../components/ScholarshipPreview';
 
 export async function generateMetadata({ params: { locale } }: { params: { locale: string } }) {
   return await getMetadata({ locale, node: graduateScholarship });
