+++ app/[locale]/academics/undergraduate/course-changes/page.tsx
@@ -2,7 +2,7 @@ import { getCourseChanges } from '@/apis/v1/academics/[type]/course-changes';
 import { getMetadata } from '@/utils/metadata';
 import { undergraduateCourseChanges } from '@/utils/segmentNode';
 
-import CourseChangesPageContent from '../../helper/course-changes/CourseChangesPageContent';
+import CourseChangesPageContent from '../../components/course-changes/CourseChangesPageContent';
 
 export async function generateMetadata({ params: { locale } }: { params: { locale: string } }) {
   return await getMetadata({ locale, node: undergraduateCourseChanges });
