+++ app/[locale]/academics/undergraduate/curriculum/create/page.tsx
@@ -5,7 +5,7 @@ import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { getPath } from '@/utils/page';
 import { curriculum } from '@/utils/segmentNode';
 
-import TimelineEditor from '../../../helper/timeline/TimelineEditor';
+import TimelineEditor from '../../../components/timeline/TimelineEditor';
 
 const curriculumPath = getPath(curriculum);
