+++ app/[locale]/academics/undergraduate/curriculum/edit/CurriculumEditPageContent.tsx
@@ -4,7 +4,7 @@ import { Curriculum } from '@/types/academics';
 import { getPath } from '@/utils/page';
 import { curriculum } from '@/utils/segmentNode';
 
-import TimelineEditor from '../../../helper/timeline/TimelineEditor';
+import TimelineEditor from '../../../components/timeline/TimelineEditor';
 
 const curriculumPath = getPath(curriculum);
