+++ app/[locale]/academics/undergraduate/general-studies-requirements/GeneralStudiesPageContent.tsx
@@ -4,7 +4,7 @@ import { deleteGeneralStudiesAction } from '@/actions/academics';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { GeneralStudiesRequirement } from '@/types/academics';
 
-import TimelineViewer from '../../helper/timeline/TimelineViewer';
+import TimelineViewer from '../../components/timeline/TimelineViewer';
 
 export default function GeneralStudiesPageContent({ data }: { data: GeneralStudiesRequirement[] }) {
   return (
