+++ app/[locale]/academics/undergraduate/general-studies-requirements/edit/GeneralStudiesEditPageContent.tsx
@@ -4,7 +4,7 @@ import { GeneralStudiesRequirement } from '@/types/academics';
 import { getPath } from '@/utils/page';
 import { curriculum } from '@/utils/segmentNode';
 
-import TimelineEditor from '../../../helper/timeline/TimelineEditor';
+import TimelineEditor from '../../../components/timeline/TimelineEditor';
 
 const curriculumPath = getPath(curriculum);
