+++ app/[locale]/academics/undergraduate/guide/edit/page.tsx
@@ -1,6 +1,6 @@
 import { getAcademicsGuide } from '@/apis/v1/academics/[type]/guide';
 
-import GuideEditPageContent from '../../../helper/guide/GuideEditPageContent';
+import GuideEditPageContent from '../../../components/guide/GuideEditPageContent';
 
 export default async function UndergraduateGuideEditPage() {
   const data = await getAcademicsGuide('undergraduate');
