+++ app/[locale]/academics/undergraduate/scholarship/create/page.tsx
@@ -1,4 +1,4 @@
-import ScholarshipCreatePage from '../../../helper/ScholarshipCreatePage';
+import ScholarshipCreatePage from '../../../components/ScholarshipCreatePage';
 
 export default function UndergraduateScholarshipCreatePage() {
   return <ScholarshipCreatePage type="undergraduate" />;
